main fragment refactoring
