+++ app/src/main/java/com/example/controle/ui/AddPersonFragment.kt
@@ -15,6 +15,8 @@ import com.example.controle.R
 import com.example.controle.animation.Effects
 import com.example.controle.dao.ProductDatabase
 import com.example.controle.model.Person
+import com.example.controle.util.AjudaFragment
+import com.example.controle.util.BaseFragment
 import com.github.rtoshiro.util.format.SimpleMaskFormatter
 import com.github.rtoshiro.util.format.text.MaskTextWatcher
 import kotlinx.android.synthetic.main.fragment_add_person.*
