+++ app/src/main/java/com/example/controle/util/DateUtils.kt
@@ -1,15 +1,11 @@
 package com.example.controle.util
 
-import android.content.Context
 import android.os.Build
-import android.util.Log
 import androidx.annotation.RequiresApi
 import com.example.controle.model.Consumer
 import com.example.controle.model.Expenses
 import com.example.controle.model.Product
-import com.example.controle.ui.BaseFragment
 import java.text.SimpleDateFormat
-import java.time.LocalDateTime
 import java.util.*
 import kotlin.math.absoluteValue
