+++ app/src/main/res/layout/fragment_ajuda.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.AjudaFragment">
+    tools:context=".util.AjudaFragment">
 
 
     <com.github.barteksc.pdfviewer.PDFView
